+++ paddle/phi/kernels/cpu/norm_kernel.cc
@@ -52,8 +52,8 @@ void NormKernel(const Context& dev_ctx,
 
   auto* place = dev_ctx.eigen_device();
 
-  Eigen::DSizes<int, 3> shape(pre, n, post);
-  Eigen::DSizes<int, 2> norm_shape(pre, post);
+  Eigen::DSizes<int64_t, 3> shape(pre, n, post);
+  Eigen::DSizes<int64_t, 2> norm_shape(pre, post);
 
   auto x_e = phi::EigenVector<T>::Flatten(x);
   auto y_e = phi::EigenVector<T>::Flatten(*out);
